+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 * Changlog
+* Str_extension validation rule
 
 ## [1.1.0] - 2019-03-03
 ### Added